refactor: Вторая доработка по замечаниям наставника
- Перенесет footer.tsx в отдельную директорию.
- Избавится от принудительного роутинга на таб Overview, сделает записьв Outlet по умолчанию при переходе к карточке фильма из адресной строки
- Уберет сценарий eject из package.json
