+++ project/package.json
@@ -16,7 +16,6 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test --passWithNoTests",
-    "eject": "react-scripts eject",
     "lint": "eslint --ext .tsx --ext .ts src/"
   },
   "eslintConfig": {
