+++ project/src/pages/my-list/my-list.tsx
@@ -1,6 +1,6 @@
 import FilmCards from '../../components/film-cards/film-cards';
 import {Films} from '../../mock/films';
-import Footer from '../../components/footer';
+import Footer from '../../components/footer/footer';
 import Logo from '../../components/header/logo';
 import UserBlock from '../../components/header/user-block';
